refactor: move all middleware to subdirectory
