+++ providers/AppProvider.ts
@@ -17,7 +17,7 @@ import { requireAll } from '@poppinss/utils'
 import { ApplicationContract } from '@poppinss/application'
 import { Emitter } from '@poppinss/events'
 import { Server, HttpContext, MiddlewareStore, Router, routePreProcessor } from '@poppinss/http-server'
-import { Cors } from '../src/Cors'
+import { Cors } from '../src/Middleware/Cors'
 
 import { HttpExceptionHandler } from '../src/HttpExceptionHandler'
 import { envLoader } from '../src/envLoader'
@@ -131,9 +131,9 @@ export default class AppProvider {
    * Registers `cors` middleware to the container
    */
   protected $registerCorsMiddleware () {
-    this.$container.bind('Adonis/Core/CorsMiddleware', () => {
-      const Config = this.$container.use('Adonis/Core/Config')
-      return new Cors(Config.get('cors', {}))
+    this.$container.singleton('Adonis/Core/CorsMiddleware', () => {
+      const config = this.$container.use('Adonis/Core/Config').get('cors', {})
+      return new Cors(config)
     })
   }
